include LICENSE file in gem
By including a copy of the exact MIT license text in the deep_merge gem,
we make the terms of the license very clear.

Also, in the specific case of deep_merge's MIT license, all derivative
copies of deep_merge are required to contain the disclaimer text within
the MIT license (as described in the sentence "The above copyright
notice and this permission notice shall be included in all copies or
substantial portions of the Software.") So in order to ease development
and distribution of deep_merge, it makes sense to simply include this
text in the gem.