Remove hash elements that have been "knocked out" with the knockout prefix
The readme for this projects thats that "if the `knockout_prefix` is passed alone as a string, it will cause the entire element to be removed". This, however, is not correct. In reality, if the `knockout_prefix` is passed alone as a string then the elemdent is set to an empty string, rather than being removed from the destination hash.